Update build targets to default to v1 instead of v1beta2, as v1 is now the current version. Also update cached API discovery for v1beta2. Update dependent APIs for targets that directly depend on v1beta2 differences. 	Change on 2014/06/03 by yarbrough <yarbrough@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=68484050

+++ gcs/src/main/java/com/google/cloud/hadoop/gcsio/ObjectWriteConditions.java
@@ -19,8 +19,6 @@ package com.google.cloud.hadoop.gcsio;
 import com.google.api.services.storage.Storage.Objects.Insert;
 import com.google.common.base.Optional;
 
-import java.math.BigInteger;
-
 /**
  * Conditions on which a object write should be allowed to continue. Corresponds to
  * setting IfGenerationmatch and IfMetaGenerationMatch in API requests.
@@ -67,12 +65,11 @@ public class ObjectWriteConditions {
    */
   public void apply(Insert objectToInsert) {
     if (hasContentGenerationMatch()) {
-      objectToInsert.setIfGenerationMatch(BigInteger.valueOf(getContentGenerationMatch()));
+      objectToInsert.setIfGenerationMatch(getContentGenerationMatch());
     }
 
     if (hasMetaGenerationMatch()) {
-      objectToInsert.setIfMetagenerationMatch(
-          BigInteger.valueOf(getMetaGenerationMatch()));
+      objectToInsert.setIfMetagenerationMatch(getMetaGenerationMatch());
     }
   }
 }
